Dashboard Service done!

+++ backend/src/main/java/org/aryan/backend/model/dto/DashBoardDTO.java
@@ -8,7 +8,7 @@ public record DashBoardDTO(
         int operationalMonitors,
         int degradedMonitors,
         int stormAlert,
-        int overallUptime,
+        double overallUptime,
         String degradedMessage,
         String stormMessage,
